Add ContentMerger. Fix tests.
