+++ LeanplumSDK/LeanplumSDK/Classes/Features/Actions/LPActionContext.m
@@ -378,7 +378,7 @@ typedef void (^LPFileCallback)(NSString* value, NSString *defaultValue);
     }
 
     NSDictionary *defaultArgs = [[[ActionManager shared] definitionWithName:actionArgs[LP_VALUE_ACTION_ARG]] values];
-    actionArgs = [[ActionManager shared] mergeWithVars:defaultArgs diff:actionArgs];
+    actionArgs = [[ContentMerger shared] mergeWithVars:defaultArgs diff:actionArgs];
     
     return actionArgs;
     LP_END_TRY
